Fix style checks.
